% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elephant102}
\alias{elephant102}
\title{Elephant gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("yeast", version = 102)}
}
\usage{
elephant102
}
\description{
Elephant gene annotations from Ensembl release 102
}
\examples{
elephant102
group_by(elephant102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
