% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elephant104}
\alias{elephant104}
\title{Elephant gene annotations from Ensembl release 104}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("yeast", version = 104)}
}
\usage{
elephant104
}
\description{
Elephant gene annotations from Ensembl release 104
}
\examples{
elephant104
group_by(elephant104, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
