% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elephant112}
\alias{elephant112}
\title{Elephant gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("yeast", version = 112)}
}
\usage{
elephant112
}
\description{
Elephant gene annotations from Ensembl release 112
}
\examples{
elephant112
group_by(elephant112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
