% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elephant98}
\alias{elephant98}
\title{Elephant gene annotations from Ensembl release 98}
\format{A tibble with 23,245 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("yeast", version = 98)}
}
\usage{
elephant98
}
\description{
Elephant gene annotations from Ensembl release 98
}
\examples{
elephant98
group_by(elephant98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
