% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly100}
\alias{fly100}
\title{Fruitfly gene annotations from Ensembl release 100}
\format{A tibble with 17,753 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("fly", version = 100)}
}
\usage{
fly100
}
\description{
Fruitfly gene annotations from Ensembl release 100
}
\examples{
fly100
group_by(fly100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
