% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly102}
\alias{fly102}
\title{Fruitfly gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("fly", version = 102)}
}
\usage{
fly102
}
\description{
Fruitfly gene annotations from Ensembl release 102
}
\examples{
fly102
group_by(fly102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
