% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly112}
\alias{fly112}
\title{Fruitfly gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("fly", version = 112)}
}
\usage{
fly112
}
\description{
Fruitfly gene annotations from Ensembl release 112
}
\examples{
fly112
group_by(fly112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
