% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly90}
\alias{fly90}
\title{Fruitfly gene annotations from Ensembl release 90}
\format{A tibble with 17,559 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("fly", version = 90)}
}
\usage{
fly90
}
\description{
Fruitfly gene annotations from Ensembl release 90
}
\examples{
data(fly90)
fly90
group_by(fly90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
