% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly92}
\alias{fly92}
\title{Fruitfly gene annotations from Ensembl release 92}
\format{A tibble with 17,737 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("fly", version = 92)}
}
\usage{
fly92
}
\description{
Fruitfly gene annotations from Ensembl release 92
}
\examples{
fly92
inner_join( group_by(fly90, biotype) \%>\% summarize(n90=n()),
            group_by(fly92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
