% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly98}
\alias{fly98}
\title{Fruitfly gene annotations from Ensembl release 98}
\format{A tibble with 17,753 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("fly", version = 98)}
}
\usage{
fly98
}
\description{
Fruitfly gene annotations from Ensembl release 98
}
\examples{
fly98
group_by(fly98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
