% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hciRdata-package.R
\docType{package}
\name{hciRdata}
\alias{hciRdata-package}
\alias{hciRdata}
\title{Datasets for differential expression and pathway analyses at HCI}
\description{
The feature annotations for the latest Ensembl 98 release are in human98,
mouse98, elephant98, fly98, pig98, rabbit98, rat98, sheep98, vervet98,
worm98, yeast98 and zebrafish98
}
