% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human100}
\alias{human100}
\title{Human gene annotations from Ensembl release 100}
\format{A tibble with 60,623 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("human", version = 100)}
}
\usage{
human100
}
\description{
Human gene annotations from Ensembl release 100
}
\examples{
human100
attr(human100, "downloaded")
group_by(human100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
