% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human102}
\alias{human102}
\title{Human gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("human", version = 102)}
}
\usage{
human102
}
\description{
Human gene annotations from Ensembl release 102
}
\examples{
human102
attr(human102, "downloaded")
group_by(human102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
