% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human104}
\alias{human104}
\title{Human gene annotations from Ensembl release 104}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("human", version = 104)}
}
\usage{
human104
}
\description{
Human gene annotations from Ensembl release 104
}
\examples{
human104
attr(human104, "downloaded")
group_by(human104, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
