% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human90}
\alias{human90}
\title{Human gene annotations from Ensembl release 90}
\format{A tibble with 63,967 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("human", version = 90)}
}
\usage{
human90
}
\description{
Human gene annotations from Ensembl release 90
}
\examples{
data(human90)
human90
attr(human90, "downloaded")
group_by(human90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
