% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human92}
\alias{human92}
\title{Human gene annotations from Ensembl release 92}
\format{A tibble with 58,395 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("human", version = 92)}
}
\usage{
human92
}
\description{
Human gene annotations from Ensembl release 92
}
\examples{
human92
attr(human92, "downloaded")
inner_join( group_by(human90, biotype) \%>\% summarize(n90=n()),
            group_by(human92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
