% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human98}
\alias{human98}
\title{Human gene annotations from Ensembl release 98}
\format{A tibble with 60,623 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("human", version = 98)}
}
\usage{
human98
}
\description{
Human gene annotations from Ensembl release 98
}
\examples{
human98
attr(human98, "downloaded")
group_by(human98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
