% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse100}
\alias{mouse100}
\title{Mouse gene annotations from Ensembl release 100}
\format{A tibble with 55,421 rows and 10 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts, human_homolog from MGI 6.14.}
\source{
\code{read_biomart("mouse", version = 100)}.
 the \code{mouse.R} file in the \code{data-raw} directory for details.
}
\usage{
mouse100
}
\description{
Mouse gene annotations from Ensembl release 100
}
\examples{
mouse100
filter(mouse100[, c(1:4,10)], grepl(",", human_homolog))
group_by(mouse100, biotype) \%>\%
 summarize(n=n(), human_homologs = sum(!is.na(human_homolog))) \%>\%
  arrange(desc(n))
}
\keyword{datasets}
