% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse104}
\alias{mouse104}
\title{Mouse gene annotations from Ensembl release 104}
\format{
A tibble with 10 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts, human_homolog from MGI 6.14.
}
\source{
\code{read_biomart("mouse", version = 104)}.
 the \code{mouse.R} file in the \code{data-raw} directory for details.
}
\usage{
mouse104
}
\description{
Mouse gene annotations from Ensembl release 104
}
\examples{
mouse104
filter(mouse104[, c(1:4,10)], grepl(",", human_homolog))
group_by(mouse104, biotype) \%>\%
 summarize(n=n(), human_homologs = sum(!is.na(human_homolog))) \%>\%
  arrange(desc(n))
}
\keyword{datasets}
