% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse90}
\alias{mouse90}
\title{Mouse gene annotations from Ensembl release 90}
\format{A tibble with 53,110 rows and 10 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts, human_homolog from MGI.}
\source{
\code{read_biomart_version("mouse", version = 90)}.
 the \code{mouse.R} file in the \code{data-raw} directory for details.
}
\usage{
mouse90
}
\description{
Mouse gene annotations from Ensembl release 90
}
\examples{
data(mouse90)
mouse90
filter(mouse90[, c(1:4,10)], grepl(",", human_homolog))
group_by(mouse90, biotype) \%>\%
 summarize(n=n(), human_homologs = sum(!is.na(human_homolog))) \%>\%
  arrange(desc(n))
}
\keyword{datasets}
