% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse92_homologs}
\alias{mouse92_homologs}
\title{Mouse and human homologs from Ensembl release 92}
\format{A tibble with 24,670 rows and 5 columns: id,mouse, perc_id, n, human}
\source{
\code{read_biomart}. See \code{data_raw/ensembl.R} for details.
}
\usage{
mouse92_homologs
}
\description{
Mouse and human homologs from Ensembl release 92
}
\examples{
mouse92_homologs
filter(mouse92_homologs, !mouse \%in\% mgi$mouse)
}
\keyword{datasets}
