% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_cancer}
\alias{msig_cancer}
\title{C4 cancer sets in the MSigDB.}
\format{A list of 5 lists including CAR (Novartis carcinoma compendium),
GCM (Global Cancer Map), GNF2 (Human tissue compendium),  MODULE (Gene sets from Segal 2004),
and MORF (unpublished compendium).  See \code{data_raw/msigdb.R} for details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig_cancer
}
\description{
C4 cancer sets in the MSigDB.
}
\examples{
sapply(msig_cancer, length)
msig_cancer$CAR[1:3]
}
\keyword{datasets}
