% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_celltype}
\alias{msig_celltype}
\title{C8 cell type signature gene sets in the MSigDB v2023.2.}
\format{
A list with cell type signatures.  See \code{data_raw/msigdb.R} for details.
}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig_celltype
}
\description{
C8 cell type signature gene sets in the MSigDB v2023.2.
}
\examples{
length(msig_celltype)
msig_celltype[1:3]
}
\keyword{datasets}
