% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_go}
\alias{msig_go}
\title{C5 Gene Ontology sets in the MSigDB.}
\format{A list of 3 lists with BP, MF, and CC.  See \code{data_raw/msigdb.R} for details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig_go
}
\description{
C5 Gene Ontology sets in the MSigDB.
}
\examples{
sapply(msig_go, length)
msig_go$BP[1:3]
}
\keyword{datasets}
