% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_hallmark}
\alias{msig_hallmark}
\title{Hallmark gene sets in MSigDB.}
\format{A list with 50 gene sets.  See \code{data_raw/msigdb.R} for details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig_hallmark
}
\description{
Hallmark gene sets in MSigDB.
}
\examples{
data(msig_hallmark)
length(msig_hallmark)
msig_hallmark[1:3]
}
\keyword{datasets}
