% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_motifs}
\alias{msig_motifs}
\title{C3 motifs in the MSigDB.}
\format{A list of 2 lists including transcription factor and miRNA targets.
See \code{data_raw/msigdb.R} for details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig_motifs
}
\description{
C3 motifs in the MSigDB.
}
\examples{
sapply(msig_motifs, length)
msig_motifs$MIR[1:3]
}
\keyword{datasets}
