% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_oncogenic}
\alias{msig_oncogenic}
\title{C6 oncogenic signatures in the MSigDB.}
\format{A list with 189 gene sets.  See \code{data_raw/msigdb.R} for details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig_oncogenic
}
\description{
C6 oncogenic signatures in the MSigDB.
}
\examples{
length(msig_oncogenic)
msig_oncogenic[1:3]
}
\keyword{datasets}
