% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pig100}
\alias{pig100}
\title{Pig gene annotations from Ensembl release 100}
\format{A tibble with 31,907 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("pig", version = 100)}
}
\usage{
pig100
}
\description{
Pig gene annotations from Ensembl release 100
}
\examples{
pig100
group_by(pig100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
