% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pig102}
\alias{pig102}
\title{Pig gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("pig", version = 102)}
}
\usage{
pig102
}
\description{
Pig gene annotations from Ensembl release 102
}
\examples{
pig102
group_by(pig102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
