% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pig112}
\alias{pig112}
\title{Pig gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("pig", version = 112)}
}
\usage{
pig112
}
\description{
Pig gene annotations from Ensembl release 112
}
\examples{
pig112
group_by(pig112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
