% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pig90}
\alias{pig90}
\title{Pig gene annotations from Ensembl release 90}
\format{A tibble with 25,880 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("pig", version = 90)}
}
\usage{
pig90
}
\description{
Pig gene annotations from Ensembl release 90
}
\examples{
data(pig90)
pig90
group_by(pig90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
