% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pig92}
\alias{pig92}
\title{Pig gene annotations from Ensembl release 92}
\format{A tibble with 25,880 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("pig", version = 92)}
}
\usage{
pig92
}
\description{
Pig gene annotations from Ensembl release 92
}
\examples{
data(pig92)
pig92
data(pig90)
inner_join( group_by(pig90, biotype) \%>\% summarize(n90=n()),
            group_by(pig92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
