% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pig98}
\alias{pig98}
\title{Pig gene annotations from Ensembl release 98}
\format{A tibble with 31,907 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("pig", version = 98)}
}
\usage{
pig98
}
\description{
Pig gene annotations from Ensembl release 98
}
\examples{
pig98
group_by(pig98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
