% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit100}
\alias{rabbit100}
\title{Rabbit gene annotations from Ensembl release 100}
\format{A tibble with 29,587 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("rabbit", version = 100)}
}
\usage{
rabbit100
}
\description{
Rabbit gene annotations from Ensembl release 100
}
\examples{
rabbit100
group_by(rabbit100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
