% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit102}
\alias{rabbit102}
\title{Rabbit gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("rabbit", version = 102)}
}
\usage{
rabbit102
}
\description{
Rabbit gene annotations from Ensembl release 102
}
\examples{
rabbit102
group_by(rabbit102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
