% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit104}
\alias{rabbit104}
\title{Rabbit gene annotations from Ensembl release 104}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("rabbit", version = 104)}
}
\usage{
rabbit104
}
\description{
Rabbit gene annotations from Ensembl release 104
}
\examples{
rabbit104
group_by(rabbit104, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
