% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit112}
\alias{rabbit112}
\title{Rabbit gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("rabbit", version = 112)}
}
\usage{
rabbit112
}
\description{
Rabbit gene annotations from Ensembl release 112
}
\examples{
rabbit112
group_by(rabbit112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
