% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit90}
\alias{rabbit90}
\title{Rabbit gene annotations from Ensembl release 90}
\format{A tibble with 23,669 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("rabbit", version = 90)}
}
\usage{
rabbit90
}
\description{
Rabbit gene annotations from Ensembl release 90
}
\examples{
rabbit90
group_by(rabbit90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
