% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit92}
\alias{rabbit92}
\title{Rabbit gene annotations from Ensembl release 92}
\format{A tibble with 23,669 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("rabbit", version = 92)}
}
\usage{
rabbit92
}
\description{
Rabbit gene annotations from Ensembl release 92
}
\examples{
rabbit92
inner_join( group_by(rabbit90, biotype) \%>\% summarize(n90=n()),
            group_by(rabbit92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
