% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rabbit98}
\alias{rabbit98}
\title{Rabbit gene annotations from Ensembl release 98}
\format{A tibble with 29,587 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("rabbit", version = 98)}
}
\usage{
rabbit98
}
\description{
Rabbit gene annotations from Ensembl release 98
}
\examples{
rabbit98
group_by(rabbit98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
