% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat100}
\alias{rat100}
\title{Rat gene annotations from Ensembl release 100}
\format{A tibble with 32,883 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("rat", version = 100)}
}
\usage{
rat100
}
\description{
Rat gene annotations from Ensembl release 100
}
\examples{
rat100
group_by(rat100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
