% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat102}
\alias{rat102}
\title{Rat gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("rat", version = 102)}
}
\usage{
rat102
}
\description{
Rat gene annotations from Ensembl release 102
}
\examples{
rat102
group_by(rat102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
