% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat104}
\alias{rat104}
\title{Rat gene annotations from Ensembl release 104}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("rat", version = 104)}
}
\usage{
rat104
}
\description{
Rat gene annotations from Ensembl release 104
}
\examples{
rat104
group_by(rat104, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
