% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat112}
\alias{rat112}
\title{Rat gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("rat", version = 112)}
}
\usage{
rat112
}
\description{
Rat gene annotations from Ensembl release 112
}
\examples{
rat112
group_by(rat112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
