% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat90}
\alias{rat90}
\title{Rat gene annotations from Ensembl release 90}
\format{A tibble with 32,883 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("rat", version = 90)}
}
\usage{
rat90
}
\description{
Rat gene annotations from Ensembl release 90
}
\examples{
rat90
group_by(rat90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
