% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat92}
\alias{rat92}
\title{Rat gene annotations from Ensembl release 92}
\format{A tibble with 32,883 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("rat", version = 92)}
}
\usage{
rat92
}
\description{
Rat gene annotations from Ensembl release 92
}
\examples{
data(rat92)
rat92
data(rat90)
inner_join( group_by(rat90, biotype) \%>\% summarize(n90=n()),
            group_by(rat92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
