% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rat98}
\alias{rat98}
\title{Rat gene annotations from Ensembl release 98}
\format{A tibble with 32,883 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("rat", version = 98)}
}
\usage{
rat98
}
\description{
Rat gene annotations from Ensembl release 98
}
\examples{
rat98
group_by(rat98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
