% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheep100}
\alias{sheep100}
\title{Sheep gene annotations from Ensembl release 100}
\format{A tibble with 27,054 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("sheep", version = 100)}.
}
\usage{
sheep100
}
\description{
Sheep gene annotations from Ensembl release 100
}
\examples{
sheep100
group_by(sheep100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
