% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheep102}
\alias{sheep102}
\title{Sheep gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("sheep", version = 102)}.
}
\usage{
sheep102
}
\description{
Sheep gene annotations from Ensembl release 102
}
\examples{
sheep102
group_by(sheep102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
