% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheep112}
\alias{sheep112}
\title{Sheep gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("sheep", version = 112)}.
}
\usage{
sheep112
}
\description{
Sheep gene annotations from Ensembl release 112
}
\examples{
sheep112
group_by(sheep112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
