% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheep90}
\alias{sheep90}
\title{Sheep gene annotations from Ensembl release 90}
\format{A tibble with 27,054 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("sheep", version = 90)}
}
\usage{
sheep90
}
\description{
Sheep gene annotations from Ensembl release 90
}
\examples{
data(sheep90)
sheep90
group_by(sheep90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
