% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheep92}
\alias{sheep92}
\title{Sheep gene annotations from Ensembl release 92}
\format{A tibble with 27,054 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("sheep", version = 92)}
}
\usage{
sheep92
}
\description{
Sheep gene annotations from Ensembl release 92
}
\examples{
data(sheep92)
sheep92
data(sheep90)
inner_join( group_by(sheep90, biotype) \%>\% summarize(n90=n()),
            group_by(sheep92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
