% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheep98}
\alias{sheep98}
\title{Sheep gene annotations from Ensembl release 98}
\format{A tibble with 27,054 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("sheep", version = 98)}
}
\usage{
sheep98
}
\description{
Sheep gene annotations from Ensembl release 98
}
\examples{
sheep98
group_by(sheep98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
