% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vervet100}
\alias{vervet100}
\title{Vervet gene annotations from Ensembl release 100}
\format{A tibble with 27,1005 rows and 10 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts, and human_homolog with maximum percent
identity from Ensembl}
\source{
\code{read_biomart("vervet", version = 100)}
}
\usage{
vervet100
}
\description{
Vervet gene annotations from Ensembl release 100
}
\examples{
vervet100
}
\keyword{datasets}
