% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vervet102}
\alias{vervet102}
\title{Vervet gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("vervet", version = 102)}
}
\usage{
vervet102
}
\description{
Vervet gene annotations from Ensembl release 102
}
\examples{
vervet102
}
\keyword{datasets}
