% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vervet104}
\alias{vervet104}
\title{Vervet gene annotations from Ensembl release 104}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("vervet", version = 104)}
}
\usage{
vervet104
}
\description{
Vervet gene annotations from Ensembl release 104
}
\examples{
vervet104
}
\keyword{datasets}
