% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vervet112}
\alias{vervet112}
\title{Vervet gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("vervet", version = 112)}
}
\usage{
vervet112
}
\description{
Vervet gene annotations from Ensembl release 112
}
\examples{
vervet112
}
\keyword{datasets}
