% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vervet98}
\alias{vervet98}
\title{Vervet gene annotations from Ensembl release 98}
\format{A tibble with 27,985 rows and 10 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts, and human_homolog with maximum percent
identity from Ensembl}
\source{
\code{read_biomart("vervet", version = 98)}
}
\usage{
vervet98
}
\description{
Vervet gene annotations from Ensembl release 98
}
\examples{
vervet98
}
\keyword{datasets}
