% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worm100}
\alias{worm100}
\title{Roundworm (C. elegans) gene annotations from Ensembl release 100}
\format{A tibble with 46,904 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("roundworm", version = 100)}
}
\usage{
worm100
}
\description{
Roundworm (C. elegans) gene annotations from Ensembl release 100
}
\examples{
worm100
group_by(worm100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
