% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worm102}
\alias{worm102}
\title{Roundworm (C. elegans) gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("roundworm", version = 102)}
}
\usage{
worm102
}
\description{
Roundworm (C. elegans) gene annotations from Ensembl release 102
}
\examples{
worm102
group_by(worm102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
