% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worm112}
\alias{worm112}
\title{Roundworm (C. elegans) gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("roundworm", version = 112)}
}
\usage{
worm112
}
\description{
Roundworm (C. elegans) gene annotations from Ensembl release 112
}
\examples{
worm112
group_by(worm112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
