% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worm90}
\alias{worm90}
\title{Roundworm (C. elegans) gene annotations from Ensembl release 90}
\format{A tibble with 46,739 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("roundworm", version = 90)}
}
\usage{
worm90
}
\description{
Roundworm (C. elegans) gene annotations from Ensembl release 90
}
\examples{
worm90
group_by(worm90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
