% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worm92}
\alias{worm92}
\title{Roundworm (C. elegans) gene annotations from Ensembl release 92}
\format{A tibble with 46,778 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("roundworm", version = 92)}
}
\usage{
worm92
}
\description{
Roundworm (C. elegans) gene annotations from Ensembl release 92
}
\examples{
worm92
inner_join( group_by(worm90, biotype) \%>\% summarize(n90=n()),
            group_by(worm92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
