% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worm98}
\alias{worm98}
\title{Roundworm (C. elegans) gene annotations from Ensembl release 98}
\format{A tibble with 46,904 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("roundworm", version = 98)}
}
\usage{
worm98
}
\description{
Roundworm (C. elegans) gene annotations from Ensembl release 98
}
\examples{
worm98
group_by(worm98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
