% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast100}
\alias{yeast100}
\title{Yeast (S. cerevisiae) gene annotations from Ensembl release 100}
\format{A tibble with 7,127 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("yeast", version = 100)}. Gene names have been replaced
}
\usage{
yeast100
}
\description{
Yeast (S. cerevisiae) gene annotations from Ensembl release 100
}
\examples{
yeast100
group_by(yeast100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
