% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast102}
\alias{yeast102}
\title{Yeast (S. cerevisiae) gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("yeast", version = 102)}. Gene names have been replaced
}
\usage{
yeast102
}
\description{
Yeast (S. cerevisiae) gene annotations from Ensembl release 102
}
\examples{
yeast102
group_by(yeast102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
