% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast104}
\alias{yeast104}
\title{Yeast (S. cerevisiae) gene annotations from Ensembl release 104}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("yeast", version = 104)}. Gene names have been replaced
}
\usage{
yeast104
}
\description{
Yeast (S. cerevisiae) gene annotations from Ensembl release 104
}
\examples{
yeast104
group_by(yeast104, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
