% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast112}
\alias{yeast112}
\title{Yeast (S. cerevisiae) gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("yeast", version = 112)}. Gene names have been replaced
}
\usage{
yeast112
}
\description{
Yeast (S. cerevisiae) gene annotations from Ensembl release 112
}
\examples{
yeast112
group_by(yeast112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
