% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast90}
\alias{yeast90}
\title{Yeast (S. cerevisiae) gene annotations from Ensembl release 90}
\format{A tibble with 7,126 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("yeast", version = 90)}
}
\usage{
yeast90
}
\description{
Yeast (S. cerevisiae) gene annotations from Ensembl release 90
}
\examples{
data(yeast90)
yeast90
group_by(yeast90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
