% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast92}
\alias{yeast92}
\title{Yeast (S. cerevisiae) gene annotations from Ensembl release 92}
\format{A tibble with 7,126 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("yeast", version = 92)}
}
\usage{
yeast92
}
\description{
Yeast (S. cerevisiae) gene annotations from Ensembl release 92
}
\examples{
data(yeast92)
yeast92
data(yeast90)
inner_join( group_by(yeast90, biotype) \%>\% summarize(n90=n()),
            group_by(yeast92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
