% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast98}
\alias{yeast98}
\title{Yeast (S. cerevisiae) gene annotations from Ensembl release 98}
\format{A tibble with 7,127 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("yeast", version = 98)}. Gene names have been replaced
}
\usage{
yeast98
}
\description{
Yeast (S. cerevisiae) gene annotations from Ensembl release 98
}
\examples{
yeast98
group_by(yeast98, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
