% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zebrafish100}
\alias{zebrafish100}
\title{Zebrafish gene annotations from Ensembl release 100}
\format{A tibble with 35,117 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart("zebrafish", version = 100)}
}
\usage{
zebrafish100
}
\description{
Zebrafish gene annotations from Ensembl release 100
}
\examples{
zebrafish100
group_by(zebrafish100, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
