% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zebrafish102}
\alias{zebrafish102}
\title{Zebrafish gene annotations from Ensembl release 102}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("zebrafish", version = 102)}
}
\usage{
zebrafish102
}
\description{
Zebrafish gene annotations from Ensembl release 102
}
\examples{
zebrafish102
group_by(zebrafish102, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
