% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zebrafish112}
\alias{zebrafish112}
\title{Zebrafish gene annotations from Ensembl release 112}
\format{
A tibble with 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts
}
\source{
\code{read_biomart("zebrafish", version = 112)}
}
\usage{
zebrafish112
}
\description{
Zebrafish gene annotations from Ensembl release 112
}
\examples{
zebrafish112
group_by(zebrafish112, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
