% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zebrafish90}
\alias{zebrafish90}
\title{Zebrafish gene annotations from Ensembl release 90}
\format{A tibble with 35,117 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("zebrafish", version = 90)}
}
\usage{
zebrafish90
}
\description{
Zebrafish gene annotations from Ensembl release 90
}
\examples{
zebrafish90
group_by(zebrafish90, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
