% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zebrafish92}
\alias{zebrafish92}
\title{Zebrafish gene annotations from Ensembl release 92}
\format{A tibble with 36,622 rows and 9 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcripts}
\source{
\code{read_biomart_version("zebrafish", version = 92)}
}
\usage{
zebrafish92
}
\description{
Zebrafish gene annotations from Ensembl release 92
}
\examples{
zebrafish92
inner_join( group_by(zebrafish90, biotype) \%>\% summarize(n90=n()),
            group_by(zebrafish92, biotype) \%>\% summarize(n92=n())) \%>\%
            mutate(diff = n92 - n90) \%>\%  arrange(desc(n92))
}
\keyword{datasets}
