% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hciRdata-package.R
\name{hciRdata}
\alias{hciRdata}
\title{Datasets for differential expression and pathway analyses at HCI}
\description{
The feature annotations for the latest Ensembl 112 release are in human112,
mouse112, elephant112, fly112, pig112, rabbit112, rat112, sheep112, vervet112,
worm112, yeast112 and zebrafish112
}
