% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hugo}
\alias{hugo}
\title{Human gene synonyms from HUGO downloaded on June 13, 2024}
\format{
A tibble with 3 columns: name, synonym, type column
}
\source{
\url{https://g-a8b222.dd271.03c0.data.globus.org/pub/databases/genenames/hgnc/tsv/non_alt_loci_set.txt}
}
\usage{
hugo
}
\description{
Human gene synonyms from HUGO downloaded on June 13, 2024
}
\examples{
hugo
table(hugo$source)
}
\keyword{datasets}
