% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hugo}
\alias{hugo}
\title{Human gene synonyms from HUGO downloaded on Dec 14, 2020}
\format{
A tibble with 3 columns: name, synonym, source column
}
\source{
\url{https://www.genenames.org/}
}
\usage{
hugo
}
\description{
Human gene synonyms from HUGO downloaded on Dec 14, 2020
}
\examples{
hugo
table(hugo$source)
}
\keyword{datasets}
