% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgi}
\alias{mgi}
\title{Mouse and human homologs from MGI}
\format{A tibble with 18,117 rows and 4 variables:
\describe{
  \item{id}{ MGI ID }
  \item{mouse}{ Mouse gene symbol }
  \item{n}{ Number of human homologs}
  \item{human}{ Human gene symbol, comma-separated if more than two }
}}
\source{
\url{http://www.informatics.jax.org/downloads/reports/HMD_HumanPhenotype.rpt}.
See the \code{data-raw} directory for code to download and group by mouse gene.
}
\usage{
mgi
}
\description{
Mouse and human homologs from MGI
}
\examples{
data(mgi)
mgi
filter(mgi, n>1)                     # 2 or more homologs
prop.table(table( mgi$human == toupper(mgi$mouse))) # 87\% are the same
data(mouse92)                     # 127 missing in Ensembl version 92
filter(mgi, !mouse \%in\% mouse92$gene_name)
attr(mgi, "downloaded")
attr(mgi, "version")
}
\keyword{datasets}
