% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_immunologic}
\alias{msig_immunologic}
\title{C7 immunologic signatures in the MSigDB 7.0.}
\format{A list with 4872 gene sets.  See \code{data_raw/msigdb.R} for details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig_immunologic
}
\description{
C7 immunologic signatures in the MSigDB 7.0.
}
\examples{
length(msig_immunologic)
msig_immunologic[1:3]
}
\keyword{datasets}
