% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig_pathways}
\alias{msig_pathways}
\title{C2 pathways in the MSigDB 7.2.}
\format{
A list of 10 pathway databases.  See \code{data_raw/msigdb.R} for details.
}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig_pathways
}
\description{
C2 pathways in the MSigDB 7.2.
}
\examples{
sapply(msig_pathways, length)
msig_pathways$KEGG[1:3]
}
\keyword{datasets}
